Wrapping import in conditional error messages
We have a bootstrap function that should get all the required
submodules, but that's not too much use if the fab recipes always error
out before you ever get them. So I'm wrapping the import of the
submodules in a try: and putting up an error warning, to encourage
people to get the submodules by running the `fab bootstrap` command.